Support floating number for price
fix

+++ src/main/java/com/carrot/carrotshop/shop/Sell.java
@@ -38,7 +38,7 @@ public class Sell extends Shop {
 	@Setting
 	private Location<World> sellerChest;
 	@Setting
-	private int price;
+	private float price;
 	
 	static private String type = "Sell";
 
@@ -61,7 +61,7 @@ public class Sell extends Shop {
 		price = getPrice(sign);
 		if (price < 0)
 			throw new ExceptionInInitializerError(Lang.SHOP_PRICE);
-		int cost = ShopConfig.getNode("cost", type).getInt(0);
+		float cost = ShopConfig.getNode("cost", type).getFloat(0);
 		if (cost > 0) {
 			UniqueAccount buyerAccount = CarrotShop.getEcoService().getOrCreateAccount(player.getUniqueId()).get();
 			TransactionResult result = buyerAccount.withdraw(getCurrency(), BigDecimal.valueOf(cost), CarrotShop.getCause());
